+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Basic Supports where added ( Experimental Only )
 - Reorganization code to better support plugins in the future
 - Layer Shrink feature added ( Useful for expanding filaments or on first layer to combat elephants foot)
+- Added Lightning Infill ( Experimental Only )
 
 ### Fixes
 - Fixed issue with the optimizer removing state changes before short moves